requirements.txt
python_chess==0.18.1
numpy==1.12.1
pandas==0.19.1
requests==2.13.0
beautifulsoup4==4.6.0
luigi==2.6.1
scikit_learn==0.18.1
keras==2.0.4
tensorflow==1.1.0